Minor-ish refactoring
- Refactor comm diagnostics code into separate source and header files
- Move put-get and active message implementations into separate files
