+++ runtime/src/comm/ofi/Makefile.share
@@ -21,6 +21,9 @@ COMM_LAUNCHER_SRCS = \
 COMM_SRCS = \
 	$(COMM_LAUNCHER_SRCS) \
 	comm-ofi.c \
+	comm-ofi-am.c \
+	comm-ofi-diagnostics.c \
+	comm-ofi-put-get.c \
 
 SRCS = $(COMM_SRCS)
